Move transforms out of dataset

+++ codes/model.py
@@ -7,11 +7,7 @@ import torch.nn.functional as F
 from torch.nn.parameter import Parameter
 from torch.autograd import Variable
 
-supported_rnns = {
-    'lstm': nn.LSTM,
-    'rnn': nn.RNN,
-    'gru': nn.GRU
-}
+supported_rnns = {'lstm': nn.LSTM, 'rnn': nn.RNN, 'gru': nn.GRU}
 supported_rnns_inv = dict((v, k) for k, v in supported_rnns.items())
 
 
@@ -48,14 +44,23 @@ class InferenceBatchSoftmax(nn.Module):
 
 
 class BatchRNN(nn.Module):
-    def __init__(self, input_size, hidden_size, rnn_type=nn.LSTM, bidirectional=False, batch_norm=True):
+    def __init__(self,
+                 input_size,
+                 hidden_size,
+                 rnn_type=nn.LSTM,
+                 bidirectional=False,
+                 batch_norm=True):
         super(BatchRNN, self).__init__()
         self.input_size = input_size
         self.hidden_size = hidden_size
         self.bidirectional = bidirectional
-        self.batch_norm = SequenceWise(nn.BatchNorm1d(input_size)) if batch_norm else None
-        self.rnn = rnn_type(input_size=input_size, hidden_size=hidden_size,
-                            bidirectional=bidirectional, bias=False)
+        self.batch_norm = SequenceWise(
+            nn.BatchNorm1d(input_size)) if batch_norm else None
+        self.rnn = rnn_type(
+            input_size=input_size,
+            hidden_size=hidden_size,
+            bidirectional=bidirectional,
+            bias=False)
         self.num_directions = 2 if bidirectional else 1
 
     def flatten_parameters(self):
@@ -66,7 +71,8 @@ class BatchRNN(nn.Module):
             x = self.batch_norm(x)
         x, _ = self.rnn(x)
         if self.bidirectional:
-            x = x.view(x.size(0), x.size(1), 2, -1).sum(2).view(x.size(0), x.size(1), -1)  # (TxNxH*2) -> (TxNxH) by sum
+            x = x.view(x.size(0), x.size(1), 2, -1).sum(2).view(
+                x.size(0), x.size(1), -1)  # (TxNxH*2) -> (TxNxH) by sum
         return x
 
 
@@ -92,14 +98,17 @@ class Lookahead(nn.Module):
         seq_len = input.size(0)
         # pad the 0th dimension (T/sequence) with zeroes whose number = context
         # Once pytorch's padding functions have settled, should move to those.
-        padding = torch.zeros(self.context, *(input.size()[1:])).type_as(input.data)
+        padding = torch.zeros(self.context, *(input.size()[1:])).type_as(
+            input.data)
         x = torch.cat((input, Variable(padding)), 0)
 
         # add lookahead windows (with context+1 width) as a fourth dimension
         # for each seq-batch-feature combination
-        x = [x[i:i + self.context + 1] for i in range(seq_len)]  # TxLxNxH - sequence, context, batch, feature
+        x = [x[i:i + self.context + 1] for i in range(seq_len)
+             ]  # TxLxNxH - sequence, context, batch, feature
         x = torch.stack(x)
-        x = x.permute(0, 2, 3, 1)  # TxNxHxL - sequence, batch, feature, context
+        x = x.permute(0, 2, 3,
+                      1)  # TxNxHxL - sequence, batch, feature, context
 
         x = torch.mul(x, self.weight).sum(dim=3)
         return x
@@ -111,8 +120,14 @@ class Lookahead(nn.Module):
 
 
 class DeepSpeech(nn.Module):
-    def __init__(self, rnn_type=nn.LSTM, labels="abc", rnn_hidden_size=768, nb_layers=5, audio_conf=None,
-                 bidirectional=True, context=20):
+    def __init__(self,
+                 rnn_type=nn.LSTM,
+                 labels="abc",
+                 rnn_hidden_size=768,
+                 nb_layers=5,
+                 audio_conf=None,
+                 bidirectional=True,
+                 context=20):
         super(DeepSpeech, self).__init__()
 
         # model metadata needed for serialization/deserialization
@@ -131,13 +146,15 @@ class DeepSpeech(nn.Module):
         num_classes = len(self._labels)
 
         self.conv = nn.Sequential(
-            nn.Conv2d(1, 32, kernel_size=(41, 11), stride=(2, 2), padding=(0, 10)),
-            nn.BatchNorm2d(32),
-            nn.Hardtanh(0, 20, inplace=True),
-            nn.Conv2d(32, 32, kernel_size=(21, 11), stride=(2, 1), ),
-            nn.BatchNorm2d(32),
-            nn.Hardtanh(0, 20, inplace=True)
-        )
+            nn.Conv2d(
+                1, 32, kernel_size=(41, 11), stride=(2, 2), padding=(0, 10)),
+            nn.BatchNorm2d(32), nn.Hardtanh(0, 20, inplace=True),
+            nn.Conv2d(
+                32,
+                32,
+                kernel_size=(21, 11),
+                stride=(2, 1),
+            ), nn.BatchNorm2d(32), nn.Hardtanh(0, 20, inplace=True))
         # Based on above convolutions and spectrogram size using conv formula (W - F + 2P)/ S+1
         rnn_input_size = int(math.floor((sample_rate * window_size) / 2) + 1)
         rnn_input_size = int(math.floor(rnn_input_size - 41) / 2 + 1)
@@ -145,34 +162,38 @@ class DeepSpeech(nn.Module):
         rnn_input_size *= 32
 
         rnns = []
-        rnn = BatchRNN(input_size=rnn_input_size, hidden_size=rnn_hidden_size, rnn_type=rnn_type,
-                       bidirectional=bidirectional, batch_norm=False)
+        rnn = BatchRNN(
+            input_size=rnn_input_size,
+            hidden_size=rnn_hidden_size,
+            rnn_type=rnn_type,
+            bidirectional=bidirectional,
+            batch_norm=False)
         rnns.append(('0', rnn))
         for x in range(nb_layers - 1):
-            rnn = BatchRNN(input_size=rnn_hidden_size, hidden_size=rnn_hidden_size, rnn_type=rnn_type,
-                           bidirectional=bidirectional)
+            rnn = BatchRNN(
+                input_size=rnn_hidden_size,
+                hidden_size=rnn_hidden_size,
+                rnn_type=rnn_type,
+                bidirectional=bidirectional)
             rnns.append(('%d' % (x + 1), rnn))
         self.rnns = nn.Sequential(OrderedDict(rnns))
         self.lookahead = nn.Sequential(
             # consider adding batch norm?
             Lookahead(rnn_hidden_size, context=context),
-            nn.Hardtanh(0, 20, inplace=True)
-        ) if not bidirectional else None
+            nn.Hardtanh(0, 20, inplace=True)) if not bidirectional else None
 
         fully_connected = nn.Sequential(
             nn.BatchNorm1d(rnn_hidden_size),
-            nn.Linear(rnn_hidden_size, num_classes, bias=False)
-        )
-        self.fc = nn.Sequential(
-            SequenceWise(fully_connected),
-        )
+            nn.Linear(rnn_hidden_size, num_classes, bias=False))
+        self.fc = nn.Sequential(SequenceWise(fully_connected), )
         self.inference_softmax = InferenceBatchSoftmax()
 
     def forward(self, x):
         x = self.conv(x)
 
         sizes = x.size()
-        x = x.view(sizes[0], sizes[1] * sizes[2], sizes[3])  # Collapse feature dimension
+        x = x.view(sizes[0], sizes[1] * sizes[2],
+                   sizes[3])  # Collapse feature dimension
         x = x.transpose(1, 2).transpose(0, 1).contiguous()  # TxNxH
 
         x = self.rnns(x)
@@ -189,14 +210,21 @@ class DeepSpeech(nn.Module):
     @classmethod
     def load_model(cls, path, cuda=False):
         package = torch.load(path, map_location=lambda storage, loc: storage)
-        model = cls(rnn_hidden_size=package['hidden_size'], nb_layers=package['hidden_layers'],
-                    labels=package['labels'], audio_conf=package['audio_conf'],
-                    rnn_type=supported_rnns[package['rnn_type']], bidirectional=package.get('bidirectional', True))
+        model = cls(
+            rnn_hidden_size=package['hidden_size'],
+            nb_layers=package['hidden_layers'],
+            labels=package['labels'],
+            audio_conf=package['audio_conf'],
+            rnn_type=supported_rnns[package['rnn_type']],
+            bidirectional=package.get('bidirectional', True))
         # the blacklist parameters are params that were previous erroneously saved by the model
         # care should be taken in future versions that if batch_norm on the first rnn is required
         # that it be named something else
-        blacklist = ['rnns.0.batch_norm.module.weight', 'rnns.0.batch_norm.module.bias',
-                     'rnns.0.batch_norm.module.running_mean', 'rnns.0.batch_norm.module.running_var']
+        blacklist = [
+            'rnns.0.batch_norm.module.weight', 'rnns.0.batch_norm.module.bias',
+            'rnns.0.batch_norm.module.running_mean',
+            'rnns.0.batch_norm.module.running_var'
+        ]
         for x in blacklist:
             if x in package['state_dict']:
                 del package['state_dict'][x]
@@ -209,28 +237,48 @@ class DeepSpeech(nn.Module):
 
     @classmethod
     def load_model_package(cls, package, cuda=False):
-        model = cls(rnn_hidden_size=package['hidden_size'], nb_layers=package['hidden_layers'],
-                    labels=package['labels'], audio_conf=package['audio_conf'],
-                    rnn_type=supported_rnns[package['rnn_type']], bidirectional=package.get('bidirectional', True))
+        model = cls(
+            rnn_hidden_size=package['hidden_size'],
+            nb_layers=package['hidden_layers'],
+            labels=package['labels'],
+            audio_conf=package['audio_conf'],
+            rnn_type=supported_rnns[package['rnn_type']],
+            bidirectional=package.get('bidirectional', True))
         model.load_state_dict(package['state_dict'])
         if cuda:
             model = torch.nn.DataParallel(model).cuda()
         return model
 
     @staticmethod
-    def serialize(model, optimizer=None, epoch=None, iteration=None, loss_results=None,
-                  cer_results=None, wer_results=None, avg_loss=None, meta=None):
+    def serialize(model,
+                  optimizer=None,
+                  epoch=None,
+                  iteration=None,
+                  loss_results=None,
+                  cer_results=None,
+                  wer_results=None,
+                  avg_loss=None,
+                  meta=None):
         model_is_cuda = next(model.parameters()).is_cuda
         model = model.module if model_is_cuda else model
         package = {
-            'version': model._version,
-            'hidden_size': model._hidden_size,
-            'hidden_layers': model._hidden_layers,
-            'rnn_type': supported_rnns_inv.get(model._rnn_type, model._rnn_type.__name__.lower()),
-            'audio_conf': model._audio_conf,
-            'labels': model._labels,
-            'state_dict': model.state_dict(),
-            'bidirectional': model._bidirectional
+            'version':
+            model._version,
+            'hidden_size':
+            model._hidden_size,
+            'hidden_layers':
+            model._hidden_layers,
+            'rnn_type':
+            supported_rnns_inv.get(model._rnn_type,
+                                   model._rnn_type.__name__.lower()),
+            'audio_conf':
+            model._audio_conf,
+            'labels':
+            model._labels,
+            'state_dict':
+            model.state_dict(),
+            'bidirectional':
+            model._bidirectional
         }
         if optimizer is not None:
             package['optim_dict'] = optimizer.state_dict()
@@ -285,11 +333,15 @@ if __name__ == '__main__':
     import os.path
     import argparse
 
-    parser = argparse.ArgumentParser(description='DeepSpeech model information')
-    parser.add_argument('--model-path', default='models/deepspeech_final.pth',
-                        help='Path to model file created by training')
+    parser = argparse.ArgumentParser(
+        description='DeepSpeech model information')
+    parser.add_argument(
+        '--model-path',
+        default='models/deepspeech_final.pth',
+        help='Path to model file created by training')
     args = parser.parse_args()
-    package = torch.load(args.model_path, map_location=lambda storage, loc: storage)
+    package = torch.load(
+        args.model_path, map_location=lambda storage, loc: storage)
     model = DeepSpeech.load_model(args.model_path)
 
     print("Model name:         ", os.path.basename(args.model_path))
@@ -306,16 +358,20 @@ if __name__ == '__main__':
     print("  Sample Rate:      ", model._audio_conf.get("sample_rate", "n/a"))
     print("  Window Type:      ", model._audio_conf.get("window", "n/a"))
     print("  Window Size:      ", model._audio_conf.get("window_size", "n/a"))
-    print("  Window Stride:    ", model._audio_conf.get("window_stride", "n/a"))
+    print("  Window Stride:    ", model._audio_conf.get(
+        "window_stride", "n/a"))
 
     if package.get('loss_results', None) is not None:
         print("")
         print("Training Information")
         epochs = package['epoch']
         print("  Epochs:           ", epochs)
-        print("  Current Loss:      {0:.3f}".format(package['loss_results'][epochs - 1]))
-        print("  Current CER:       {0:.3f}".format(package['cer_results'][epochs - 1]))
-        print("  Current WER:       {0:.3f}".format(package['wer_results'][epochs - 1]))
+        print("  Current Loss:      {0:.3f}".format(
+            package['loss_results'][epochs - 1]))
+        print("  Current CER:       {0:.3f}".format(
+            package['cer_results'][epochs - 1]))
+        print("  Current WER:       {0:.3f}".format(
+            package['wer_results'][epochs - 1]))
 
     if package.get('meta', None) is not None:
         print("")
